refactor: split out ResourceRef resolving code
The resolving code was getting a bit too many ways of enabling/
disabling the different parts of the resolver, so a new abstraction
was created for a pluggable resolver: the `ResourceResolver` and
a bunch of implementations.
