+++ src/main/java/dev/jbang/source/resolvers/StdinScriptResourceResolver.java
@@ -1,0 +1,59 @@
+package dev.jbang.source.resolvers;
+
+import static dev.jbang.util.Util.getMainClass;
+import static dev.jbang.util.Util.hasMainMethod;
+
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+import java.util.stream.Collectors;
+
+import dev.jbang.Cache;
+import dev.jbang.Settings;
+import dev.jbang.cli.BaseCommand;
+import dev.jbang.cli.ExitException;
+import dev.jbang.source.ResourceRef;
+import dev.jbang.source.ResourceResolver;
+import dev.jbang.util.Util;
+
+/**
+ * A <code>ResourceResolver</code> that, when given "-" as a resource string,
+ * will try to create a copy in the cache with a proper file name and return a
+ * reference to that file.
+ */
+public class StdinScriptResourceResolver implements ResourceResolver {
+	@Override
+	public ResourceRef resolve(String resource) {
+		ResourceRef result = null;
+
+		try {
+			// support stdin
+			if (ResourceRef.isStdin(resource)) {
+				String scriptText = new BufferedReader(
+						new InputStreamReader(System.in, StandardCharsets.UTF_8))
+																					.lines()
+																					.collect(Collectors.joining(
+																							System.lineSeparator()));
+
+				String urlHash = Util.getStableID(scriptText);
+				File cache = Settings.getCacheDir(Cache.CacheClass.stdins).resolve(urlHash).toFile();
+				cache.mkdirs();
+				String basename = urlHash;
+				String suffix = ".jsh";
+				if (hasMainMethod(scriptText)) {
+					suffix = ".java";
+					basename = getMainClass(scriptText).orElse(basename);
+				}
+				File scriptFile = new File(cache, basename + suffix);
+				Util.writeString(scriptFile.toPath(), scriptText);
+				result = ResourceRef.forCachedResource(resource, scriptFile);
+			}
+		} catch (IOException e) {
+			throw new ExitException(BaseCommand.EXIT_UNEXPECTED_STATE, "Could not cache script from stdin", e);
+		}
+
+		return result;
+	}
+}
